feat: 경주를 실행하면 자동으로 5회 실행된다.
